Moved utility functions for gaussian and hci datasets to tools module.
